Fixing extended verification code for non-windows
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1101 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/ev_root_ca_metadata.cc
@@ -184,7 +184,7 @@ EVRootCAMetadata::EVRootCAMetadata() {
   // Constructs the object from the raw metadata in ev_root_ca_metadata.
   num_policy_oids_ = arraysize(ev_root_ca_metadata);
   policy_oids_.reset(new const char*[num_policy_oids_]);
-  for (int i = 0; i < arraysize(ev_root_ca_metadata); i++) {
+  for (size_t i = 0; i < arraysize(ev_root_ca_metadata); i++) {
     const EVMetadata& metadata = ev_root_ca_metadata[i];
     ev_policy_[metadata.fingerprint] = metadata.policy_oid;
     // Multiple root CA certs may use the same EV policy OID.  Having
